Vaihe 4: Challenge-tila — 20 solo-haastetasoa tähdillä
- ChallengeLevels: kiinteä siemen per taso, vaikeus ja tähtirajat
  kiristyvät; 1 tähti = voitto, 2-3 = pisterajat
- Tähdet tallentuvat profiiliin (parasta ei huononneta), seuraava taso
  aukeaa edellisen läpäisystä
- Challenges-nappi + tasovalintadialogi tähtineen ja kokonaissaldolla
- Voittoruutu näyttää 'Level N cleared! ★★☆' ja Play again aloittaa
  saman haastetason uudelleen

+++ app/src/main/java/com/finnvek/cornersapart/model/BoardSnapshot.kt
@@ -22,7 +22,7 @@ class BoardSnapshot private constructor(
 
     init {
         require(size > 0) { "Board size must be positive." }
-        require(cells.size == size * size) { "Board cells must match board size." }
+        require(cells.size.toLong() == size.toLong() * size.toLong()) { "Board cells must match board size." }
     }
 
     override fun cellAt(index: Int): Int = cells[index]
